I will finish my own test in strmapi

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: kde-paul <kde-paul@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/01 16:07:22 by kde-paul          #+#    #+#             */
-/*   Updated: 2025/10/18 14:10:49 by kde-paul         ###   ########.fr       */
+/*   Updated: 2025/10/29 20:15:51 by kde-paul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include <libft.h>
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
